合并最新更新

+++ wzkris-modules/wzkris-modules-user/wzkris-user-biz/src/main/java/com/wzkris/user/controller/SysTenantPackageController.java
@@ -3,14 +3,19 @@ package com.wzkris.user.controller;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.wzkris.common.core.constant.CommonConstants;
 import com.wzkris.common.core.domain.Result;
+import com.wzkris.common.core.utils.BeanUtil;
 import com.wzkris.common.core.utils.StringUtil;
 import com.wzkris.common.log.annotation.OperateLog;
 import com.wzkris.common.log.enums.OperateType;
 import com.wzkris.common.orm.page.Page;
-import com.wzkris.common.security.utils.SysUtil;
+import com.wzkris.common.security.oauth2.annotation.CheckPerms;
+import com.wzkris.common.security.utils.LoginUserUtil;
 import com.wzkris.common.web.model.BaseController;
 import com.wzkris.user.domain.SysTenantPackage;
-import com.wzkris.user.domain.vo.SysMenuCheckSelectTreeVO;
+import com.wzkris.user.domain.req.EditStatusReq;
+import com.wzkris.user.domain.req.SysTenantPackageQueryReq;
+import com.wzkris.user.domain.req.SysTenantPackageReq;
+import com.wzkris.user.domain.vo.CheckedSelectTreeVO;
 import com.wzkris.user.mapper.SysTenantPackageMapper;
 import com.wzkris.user.service.SysMenuService;
 import com.wzkris.user.service.SysTenantPackageService;
@@ -34,7 +39,7 @@ import java.util.List;
 @Tag(name = "租户套餐管理")
 @Validated
 @RequiredArgsConstructor
-@PreAuthorize("@SysUtil.isSuperTenant()")// 只允许超级租户访问
+@PreAuthorize("@LoginUserUtil.isSuperTenant()")// 只允许超级租户访问
 @RestController
 @RequestMapping("/sys_tenant/package")
 public class SysTenantPackageController extends BaseController {
@@ -45,37 +50,44 @@ public class SysTenantPackageController extends BaseController {
 
     @Operation(summary = "套餐分页")
     @GetMapping("/list")
-    @PreAuthorize("@ps.hasPerms('tenant_package:list')")
-    public Result<Page<SysTenantPackage>> listPage(SysTenantPackage sysTenantPackage) {
+    @CheckPerms("tenant_package:list")
+    public Result<Page<SysTenantPackage>> listPage(SysTenantPackageQueryReq queryReq) {
         startPage();
-        List<SysTenantPackage> list = tenantPackageService.list(sysTenantPackage);
+        List<SysTenantPackage> list = tenantPackageMapper.selectList(this.buildQueryWrapper(queryReq));
         return getDataTable(list);
+    }
+
+    private LambdaQueryWrapper<SysTenantPackage> buildQueryWrapper(SysTenantPackageQueryReq queryReq) {
+        return new LambdaQueryWrapper<SysTenantPackage>()
+                .like(StringUtil.isNotNull(queryReq.getPackageName()), SysTenantPackage::getPackageName, queryReq.getPackageName())
+                .eq(StringUtil.isNotNull(queryReq.getStatus()), SysTenantPackage::getStatus, queryReq.getStatus())
+                .orderByDesc(SysTenantPackage::getPackageId);
     }
 
     @Operation(summary = "套餐选择列表(不带分页)")
     @GetMapping("/selectlist")
-    @PreAuthorize("@ps.hasPerms('tenant_package:list')")
-    public Result<List<SysTenantPackage>> selectList(SysTenantPackage sysTenantPackage) {
+    @CheckPerms(value = {"tenant:add", "tenant:edit"}, mode = CheckPerms.Mode.OR)// 租户添加修改时使用
+    public Result<List<SysTenantPackage>> selectList(String packageName) {
         LambdaQueryWrapper<SysTenantPackage> lqw = new LambdaQueryWrapper<SysTenantPackage>()
                 .select(SysTenantPackage::getPackageId, SysTenantPackage::getPackageName)
                 .eq(SysTenantPackage::getStatus, CommonConstants.STATUS_ENABLE)
-                .like(StringUtil.isNotBlank(sysTenantPackage.getPackageName()), SysTenantPackage::getPackageName, sysTenantPackage.getPackageName());
+                .like(StringUtil.isNotBlank(packageName), SysTenantPackage::getPackageName, packageName);
         return ok(tenantPackageMapper.selectList(lqw));
     }
 
     @Operation(summary = "套餐菜单选择树")
     @GetMapping({"/menu_select_tree/", "/menu_select_tree/{packageId}"})
-    @PreAuthorize("@ps.hasPerms('tenant_package:list')")
-    public Result<SysMenuCheckSelectTreeVO> tenantPackageMenuTreeList(@PathVariable(required = false) Long packageId) {
-        SysMenuCheckSelectTreeVO resp = new SysMenuCheckSelectTreeVO();
-        resp.setCheckedKeys(tenantPackageMapper.listMenuIdByPackageId(packageId));
-        resp.setMenus(menuService.listMenuSelectTree(SysUtil.getUserId()));
-        return ok(resp);
+    @CheckPerms("tenant_package:list")
+    public Result<CheckedSelectTreeVO> tenantPackageMenuTreeList(@PathVariable(required = false) Long packageId) {
+        CheckedSelectTreeVO checkedSelectTreeVO = new CheckedSelectTreeVO();
+        checkedSelectTreeVO.setCheckedKeys(tenantPackageMapper.listMenuIdByPackageId(packageId));
+        checkedSelectTreeVO.setSelectTrees(menuService.listMenuSelectTree(LoginUserUtil.getUserId()));
+        return ok(checkedSelectTreeVO);
     }
 
     @Operation(summary = "套餐详细信息")
     @GetMapping("/{packageId}")
-    @PreAuthorize("@ps.hasPerms('tenant_package:query')")
+    @CheckPerms("tenant_package:query")
     public Result<SysTenantPackage> getInfo(@NotNull(message = "[packageId] {validate.notnull}")
                                             @PathVariable Long packageId) {
         return ok(tenantPackageMapper.selectById(packageId));
@@ -84,24 +96,37 @@ public class SysTenantPackageController extends BaseController {
     @Operation(summary = "新增租户套餐")
     @OperateLog(title = "租户套餐", subTitle = "新增套餐", operateType = OperateType.INSERT)
     @PostMapping("/add")
-    @PreAuthorize("@ps.hasPerms('tenant_package:add')")
-    public Result<Void> add(@Valid @RequestBody SysTenantPackage tenantPackage) {
-        return toRes(tenantPackageMapper.insert(tenantPackage));
+    @CheckPerms("tenant_package:add")
+    public Result<Void> add(@Valid @RequestBody SysTenantPackageReq req) {
+        return toRes(tenantPackageMapper.insert(BeanUtil.convert(req, SysTenantPackage.class)));
     }
 
     @Operation(summary = "修改租户套餐")
     @OperateLog(title = "租户套餐", subTitle = "修改套餐", operateType = OperateType.UPDATE)
     @PostMapping("/edit")
-    @PreAuthorize("@ps.hasPerms('tenant_package:edit')")
-    public Result<Void> edit(@Valid @RequestBody SysTenantPackage tenantPackage) {
-        return toRes(tenantPackageMapper.updateById(tenantPackage));
+    @CheckPerms("tenant_package:edit")
+    public Result<Void> edit(@Valid @RequestBody SysTenantPackageReq req) {
+        return toRes(tenantPackageMapper.updateById(BeanUtil.convert(req, SysTenantPackage.class)));
+    }
+
+    @Operation(summary = "修改租户套餐状态")
+    @OperateLog(title = "租户套餐", subTitle = "修改租户套餐状态", operateType = OperateType.UPDATE)
+    @PostMapping("/edit_status")
+    @CheckPerms("tenant_package:edit")
+    public Result<Void> editStatus(@RequestBody @Valid EditStatusReq statusReq) {
+        SysTenantPackage update = new SysTenantPackage(statusReq.getId());
+        update.setStatus(statusReq.getStatus());
+        return toRes(tenantPackageMapper.updateById(update));
     }
 
     @Operation(summary = "删除租户套餐")
     @OperateLog(title = "租户套餐", subTitle = "删除套餐", operateType = OperateType.DELETE)
     @PostMapping("/remove")
-    @PreAuthorize("@ps.hasPerms('tenant_package:remove')")
+    @CheckPerms("tenant_package:remove")
     public Result<Void> remove(@NotEmpty(message = "[packageId] {validate.notnull}") @RequestBody List<Long> packageIds) {
+        if (tenantPackageService.checkPackageUsed(packageIds)) {
+            return fail("删除失败, 套餐正在使用");
+        }
         return toRes(tenantPackageMapper.deleteByIds(packageIds));
     }
